% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{coords_to_sf}
\alias{coords_to_sf}
\title{Converts points to a sf object}
\usage{
coords_to_sf(dat, epsg = 4326)
}
\arguments{
\item{dat}{\code{data.frame} with columns lat and lon.}

\item{epsg}{The epsg code of the output coordinate reference system.}
}
\value{
\code{sf} object of points.
}
\description{
Converts points to a sf object
}
