% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_coordinate_cube_values}
\alias{get_coordinate_cube_values}
\title{Calculates the coordinate value equivalents of a given STKDE cube}
\usage{
get_coordinate_cube_values(xyt, cube)
}
\arguments{
\item{xyt}{\code{data.frame} containing the x, y and t coordinates of the
input data (projected coordinates).}

\item{cube}{\code{arma::cube} object containing all STKDE values.}
}
\value{
\code{list} the size of the cube input dimensions filled with the
values of the xyt object.
}
\description{
Calculates the coordinate value equivalents of a given STKDE cube
}
