\name{stkde-package}
\alias{stkde-package}
\alias{stkde}
\docType{package}
\title{
\packageTitle{stkde}
}
\description{
\packageDescription{stkde}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{stkde}
\packageIndices{stkde}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{stkde}

Maintainer: \packageMaintainer{stkde}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
