% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stkde-package.r, R/stkde.R
\docType{package}
\name{stkde}
\alias{stkde}
\alias{stkde-package}
\alias{stkde}
\title{stkde.}
\usage{
stkde(xyt_in, hs, ht, x_size, y_size, t_size)
}
\arguments{
\item{xyt_in}{\code{data.frame} containing the x, y and t coordinates of the
input data (projected coordinates).}

\item{hs}{Spatial bandwidth in m.}

\item{ht}{Temporal bandwidth.}

\item{x_size}{Spatial x dimension of output array.}

\item{y_size}{Spatial y dimension of output array.}

\item{t_size}{Temporal t dimension of output array.}
}
\value{
\code{arma::cube} object containing all STKDE values.
}
\description{
stkde.

Calculates a spatiotemporal kernel density estimation (STKDE) and returns it
as a 3D-array
}
\examples{
\dontrun{
x <- runif (100) * 100
y <- runif (100) * 100
t <- runif (100) * 100
xyt <- data.frame (x, y, t)
stkde_val <- stkde (xyt_in = xyt, hs = 20, ht = 20, x_size = 15, y_size = 15,
t_size = 15)
}
}
